feat: projection para o dashboard

+++ lavaja-api/src/main/java/io/github/pietro_dev/lavajaapi/model/repository/projections/AgendamentoPorDia.java
@@ -2,5 +2,5 @@ package io.github.pietro_dev.lavajaapi.model.repository.projections;
 
 public interface AgendamentoPorDia {
     Integer getDia();
-    Integer getTotalAgendamentos();
+    Long getTotalAgendamentos();
 }
